'Add scatter plot of puff z and mass'

+++ pltPuffIds.py
@@ -3,6 +3,8 @@ import sys
 import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
+import matplotlib.cm as cm
+from matplotlib import colors
 
 def iter_loadtxt(filename, delimiter=',', skiprows=0, dtype=float):
   def iter_func():
@@ -23,48 +25,93 @@ def openFout(tEnd,tStep):
   
   fNo   = int(tEnd/tStep)
   fName = csvName.split('.')[0]+'_%d.csv'%fNo
-  print 'Creating %s'%fName
-  fOut = open(fName,'w')
-  fOut.write('%s\n'%hdr)
+  if os.path.exists(fName):
+    fOut = None
+  else:
+    print 'Creating %s'%fName
+    fOut = open(fName,'w')
+    fOut.write('%s\n'%hdr)
+
   lOpen = False
   
   return (fOut,fName,lOpen)
 
 def plotCsv(fName,tStart,tEnd,colName,colDtype):
+  
   df = pd.read_csv(fName, sep=',', skiprows=1, names=colNames, dtype=colDtype)
   #print df.columns
+  
+  print 'Time = ',df['Time'].max()
+  
+  if df['Time'].max() < 3600.:
+    return 
+  
   maxMass = df['mass'].max()
   print maxMass
-  df['massR'] = map(int,(df['mass']/maxMass)*50)
+  #df['massR'] = map(int,(df['mass']/maxMass)*50)
   
   dfNew = df[(df.Parent1==0) & (df.Routine !='ZS0')]
+  print 'Number of new puffs = ',dfNew['Self'].count()
   
-  dfzs0 = df[(df.Routine =='ZS0')]
-  #print dfzs0
+  #print df.drop_duplicates(['Routine'])['Routine']
   
+ # dfSplt = []
+ # for rName in ['ZS0']:
+ #   dfSplt = df[(df.Routine==rName)]
+ 
+ #levels = linspace(0,clrmax,num=clrlev)
+ #clrmap = cm.get_cmap('jet',clrlev-1)
+ #lnorm  = colors.Normalize(levels,clip=False)
+ 
+  clrlev = 6
+  vmin = 1e-6
+  vmax = 1.
+  levels = np.logspace(1e-5,1,num=clrlev,base=10.0)
+  clrmap = cm.get_cmap('jet',clrlev-1)
+  lnorm  = colors.LogNorm(levels,clip=False)
+   
+  for rName in ['XS0','YS0','ZS0']:
+    dfzs0 = df[(df.Routine ==rName)]
+    dfC1  = pd.merge(dfzs0,dfzs0, left_on='Child1', right_on='Self')
+    dfC1['massR'] = (dfC1['mass_y']/dfC1['mass_x'])
+    #dfC1.to_csv('dfc1_%s'%fName)
+    plt.clf()
+    cs = plt.scatter(dfC1['Time_x'],dfC1['massR'],color='red',marker='o')
+    
+    #cs = plt.scatter(dfC1['Time_x'],dfC1['zbar_x'],c=dfC1['massR'],edgecolor=None,marker='o',norm=lnorm,cmap=clrmap,\
+    #                 vmin=vmin,vmax=vmax)
+    plt.xlabel('Time(s)')
+    plt.ylabel('Z(m)')
+    plt.title('%s(%d)'%(rName,dfC1['massR'].count()))
+    #plt.colorbar(cs,ticks=[0,1],fraction=0.08)
+    plt.show()
+  
+  #print 'Number of puffs split = ',dfzs0['Child1'].index
+  '''
   dfzs1 = df[(df.Routine =='ZS1')]
   print dfzs1
   
   dfzs2 = df[(df.Routine =='ZS2')]
   print dfzs2
   
+  sys.exit()
   plt.clf()
   plt.hold(True)
   
   cA = ['red','green','blue','purple']
   mA = ['o','s','>','<']
   
-  for ityp,dtyp in enumerate([dfNew,dfzs0,dfzs1,dfzs2]):
+  for ityp,dtyp in enumerate([dfzs0,dfzs1,dfzs2]): # [dfNew,dfzs0,dfzs1,dfzs2]):
     plt.scatter(dtyp['Time'],dtyp['zbar'],color='white',alpha=0.5,edgecolors=cA[ityp],\
                 marker=mA[ityp],s=dtyp['massR'])
   
-  plt.title('Puffs for Time between %8.3f and %8.3f'%(tStart/3600.,tEnd/3600.))
+  plt.title('Puffs for Time between %8.3f and %8.3f hr'%(tStart/3600.,tEnd/3600.))
   plt.xlabel('Time(Sec)')
   plt.ylabel('Zbar(m)')
   
   plt.hold(False)
   plt.show()
-  
+  '''
   return  
 
 if sys.platform == 'win32':
@@ -74,7 +121,7 @@ else:
   
 #pData = iter_loadtxt('puffTree_3hr.csv',skiprows=1)
 
-csvName = 'puffTree_3hr.csv'
+csvName = 'x3hr.csv'
 #csvName = 'pfTree3hr_1000.csv'
 
 csvFile = open(csvName,'r')
@@ -111,19 +158,22 @@ for line in csvFile:
   else:
     colVals = line.split(',')
     time = float(colVals[0])
-    if time > 3600.:
-      break
+    #if time > 3600.:
+    #  break
     if time >= tStart and lOpen:
       fOut,fName,lOpen  = openFout(tEnd,tStep)
     if not lOpen and time > tEnd:
-      fOut.close()
+      if fOut is not None:
+        fOut.close()
       plotCsv(fName,tStart,tEnd,colNames,colDtype)
       tStart           = tEnd
       tEnd             = tStart + tStep
       fOut,fName,lOpen = openFout(tEnd,tStep)
-    fOut.write('%s\n'%line)
-
-fOut.close()    
+    if fOut is not None:
+      fOut.write('%s\n'%line)
+if fOut is not None:
+  fOut.close() 
+plotCsv(fName,tStart,tEnd,colNames,colDtype)   
 csvFile.close()
 
 sys.exit()
